Fix tests, make fields pub
